pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "em-evo-evals"
version = "0.1.0"
description = "Evaluation pipeline for measuring Emergent Misalignment in Large Language Models"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "PyYAML>=6.0",
    "openai>=1.0",
    "tenacity>=8.0",
    "pandas>=2.0",
    "matplotlib>=3.0",
    "numpy>=1.0",
    "transformers>=4.0",
    "torch>=2.0",
    "accelerate",
    "python-dotenv>=1.0",
    "vllm",
    "peft>=0.19.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
